+++ MotorControl.h
@@ -40,7 +40,7 @@ void initPins();
 __interrupt void epwm2ISR(void);
 
 // Turn motor
-void turnMotor(uint16_t motor, uint16_t direction, uint16_t degrees);
+void turnMotor(uint16_t motor, uint16_t direction, float degrees);
 
 // Enable and disable motor
 void enableMotor(uint16_t motor, uint16_t direction);
